+++ src/main/java/nl/enjarai/doabarrelroll/mixin/ClientPlayerEntityMixin.java
@@ -26,7 +26,7 @@ public abstract class ClientPlayerEntityMixin extends AbstractClientPlayerEntity
 			method = "<init>(Lnet/minecraft/client/MinecraftClient;Lnet/minecraft/client/world/ClientWorld;Lnet/minecraft/client/network/ClientPlayNetworkHandler;Lnet/minecraft/stat/StatHandler;Lnet/minecraft/client/recipebook/ClientRecipeBook;ZZ)V",
 			at = @At("RETURN")
 	)
-	public void init(MinecraftClient client, ClientWorld world, ClientPlayNetworkHandler networkHandler, StatHandler stats, ClientRecipeBook recipeBook, boolean lastSneaking, boolean lastSprinting, CallbackInfo ci) {
+	public void doABarrelRoll$init(MinecraftClient client, ClientWorld world, ClientPlayNetworkHandler networkHandler, StatHandler stats, ClientRecipeBook recipeBook, boolean lastSneaking, boolean lastSprinting, CallbackInfo ci) {
 		DoABarrelRollClient.left = ElytraMath.getAssumedLeft(getYaw());
 	}
 }